Created new folder, reformatted existing pages
Created new folder and copied the previous website into the new one.  Made some changes here:

Changed ID tags on the HTML elements for the content of the index.html and tactics.html page to class tags.

Edited the CSS to allow for dynamic re-sizing of page elements for better mobile use.

Changed the "maps.html" page to "napoleons_marshalls.html" page, changed the formatting on it to fit better.